# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenID Connect APP
  description: |-
    This is the OpenID Connect API documentation.
  termsOfService: https://localhost:8787
  contact:
    email: kotaro.otaka@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8787
tags:
  - name: OIDC
    description: OIDC
  - name: IdP
    description: IdP
paths:
  /.well-known/openid-configuration:
    get:
      tags:
        - OIDC
      summary: OpenID Provider Configuration
      description: OpenID Provider Configuration
      operationId: openIDConfiguration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenIDConfigurationResponseSchema"
        default:
          description: Internal Server Error
  /authorize:
    get:
      tags:
        - OIDC
      summary: Authorization Endpoint
      description: |
        Authorization Endpoint
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#code-authz-req
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          description: response_type
          schema:
            type: string
            enum:
              - code
        - name: client_id
          in: query
          required: true
          description: client_id
          schema:
            type: string
            format: uri
            example: http://localhost:8080
        - name: redirect_uri
          in: query
          required: false
          description: http://localhost:8080/callback
          schema:
            type: string
            format: uri
            example: http://localhost:8080/callback
        - name: scope
          in: query
          required: false
          description: openid
          schema:
            type: string
            enum:
              - openid
        - name: state
          in: query
          required: false
          description: state
          schema:
            type: string
      responses:
        "302":
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/callback
        default:
          description: Internal Server Error
  /token:
    post:
      tags:
        - OIDC
      summary: Token Endpoint
      description: |
        Token Endpoint
        ref: https://openid-foundation-japan.github.io/rfc6749.ja.html#token-req
        ref: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenRequest
      operationId: token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequestSchema'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control: 
              schema:
                type: string
                example: no-store
            Pragma:
              schema:
                type: string
                example: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseSchema'
        default:
          description: Internal Server Error
  /login:
    get:
      tags:
        - IdP
      summary: Login Veiw
      description: Login View
      operationId: loginView
      parameters:
        - name: id
          in: query
          required: true
          description: id
          schema:
            type: string
            example: xxxxxxxxxx
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
        default:
          description: Internal Server Error
    post:
      tags:
        - IdP
      summary: Login
      description: Login
      operationId: login
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: xxxxxxxxxx
                password:
                  type: string
                  example: xxxxxxxxxx
      responses:
        "302":
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8787/authorize?response_type=code&client_id=http://localhost:8080&redirect_uri=http://localhost:8080/callback&scope=openid&state=xxxxxxxxxx
        default:
          description: Internal Server Error
  /callback:
    get:
      tags:
        - IdP
      summary: Callback
      description: Callback
      operationId: callback
      parameters:
        - name: id
          in: query
          required: true
          description: id
          schema:
            type: string
            example: xxxxxxxxxx
      responses:
        "302":
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: http://localhost:8080/callback
        default:
          description: Internal Server Error
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OpenIDConfigurationResponseSchema:
      type: object
      properties:
        issuer:
          type: string
          description: http://localhost:8787
          format: uri
          example: http://localhost:8787
        authorization_endpoint:
          type: string
          description: http://localhost:8787/authorize
          format: uri
          example: http://localhost:8787/authorize
        token_endpoint:
          type: string
          description: http://localhost:8787/token
          format: uri
          example: http://localhost:8787/token
        userinfo_endpoint:
          type: string
          description: http://localhost:8787/userinfo
          format: uri
          example: http://localhost:8787/userinfo
        jwks_uri:
          type: string
          description: http://localhost:8787/certs
          format: uri
          example: http://localhost:8787/certs
        revocation_endpoint:
          type: string
          description: http://localhost:8787/revoke
          format: uri
          example: http://localhost:8787/revoke
      required:
        - issuer
        - authorization_endpoint
        - token_endpoint
        - userinfo_endpoint
        - jwks_uri
        - revocation_endpoint
    TokenRequestSchema:
      type: object
      properties:
        grant_type:
          type: string
          enum:
            - authorization_code
            - refresh_token
          description: grant_type
          example: authorization_code
        code:
          type: string
          description: code
          example: xxxxxxxxxxxxxxx
        redirect_uri:
          type: string
          description: http://localhost:8080/rp/callback
          format: uri
          example: http://localhost:8080/rp/callback
        refresh_token:
          type: string
          description: refresh_token
          example: xxxxxxxx
        client_id:
          type: string
          description: client_id
          example: client_id
        client_secret:
          type: string
          description: client_secret
          example: client_secret
        scope:
          type: string
          example: openid profile email
        code_verifier:
          type: string
          description: code_verifier
          example: code_verifier
      required:
        - grant_type
        - code
        - redirect_uri
    TokenResponseSchema:
      type: object
      description: https://openid-foundation-japan.github.io/openid-connect-core-1_0.ja.html#TokenResponse
      properties:
        access_token:
          type: string
          description: access_token
          example: access_token
        token_type:
          type: string
          description: token_type
          example: token_type
        refresh_token:
          type: string
          description: refresh_token
          example: refresh_token
        expires_in:
          type: integer
          description: expires_in
          example: 3600
        id_token:
          type: string
          description: id_token
          example: id_token
        scope:
          type: array
          description: scope
          items:
            type: string
            enum:
              - openid
              - profile
              - email
      required:
        - access_token
        - token_type
        - refresh_token
        - expires_in
        - id_token